Fix signature verification of twenty three historical cosmos-sdk/MsgEventRecord transactions on mainnet block height 8587012 through 9266259.
Transaction signatures are verified by taking the keccak hash of the transaction message serialised to a sorted, compact JSON representation and recovering the public key from that hash.

The transactions concerned all had zero length "data" fields which are serialised as "0x" in the signing representation, however the respective signatures were initially generated by serialising the zero length "data" field as "0x0" thus causing signature verification to fail and causing the transactions, and their effects on state, to be ignored.

These then leads to an app hash mismatch on the subsequent block halting replay of the chain.

For the twenty three transactions concerned as identified by chain id, block height, account number and sequence, ',"data":"0x",' is replaced with ',"data":"0x0",' in the signing representation prior to signature verification.